+++ src/components/empty-favorites-page/empty-favorites-page.tsx
