+++ src/components/empty-main-page/empty-main-page.tsx
