+++ src/components/filled-favorites-page/filled-favorites-page.tsx
@@ -1,11 +1,11 @@
-import FavoriteOfferList from '../../components/favorite-offer-list/favorite-offer-list';
+import FavoriteOfferList from '../favorite-offer-list/favorite-offer-list';
 import { Offers } from '../../mocks/mock-types/offers';
 
-type NoEmptyFavoritesPageProps = {
+type FilledFavoritesPageProps = {
   mockOffers: Offers;
 }
 
-export default function NoEmptyFavoritesPage({mockOffers}: NoEmptyFavoritesPageProps) {
+export default function FilledFavoritesPage({mockOffers}: FilledFavoritesPageProps) {
   return (
     <section className="favorites">
       <h1 className="favorites__title">Saved listing</h1>
