+++ src/components/logo/logo-link.tsx
@@ -1,5 +1,5 @@
 import { Link } from 'react-router-dom';
-import { AppRoute } from '../../data';
+import { AppRoute } from '../../utils/data';
 
 export default function LogoLink() {
   return (
