+++ src/mocks/mock-util.ts
@@ -1,3 +1,3 @@
-import { AuthorizationStatus } from '../data';
+import { AuthorizationStatus } from '../utils/data';
 
 export const isUserLogged = (logged: AuthorizationStatus) => logged === AuthorizationStatus.Auth;
