+++ src/pages/offer-page/offer-page.tsx
@@ -1,13 +1,13 @@
 import OfferImage from '../../components/offer-image/offer-image';
-import OfferReviewForm from './offer-review-form';
+import ReviewForm from '../../components/review-form/review-form';
 import NotFoundPage from '../not-found-page/not-found-page';
 import OfferInsideList from '../../components/offer-inside-list/offer-inside-list';
 import PostReviewError from '../../components/error/post-review-error';
 import { useParams } from 'react-router-dom';
-import { AuthorizationStatus } from '../../data';
+import { AuthorizationStatus } from '../../utils/data';
 import { isUserLogged } from '../../mocks/mock-util';
 import { Offer, Offers } from '../../mocks/mock-types/offers';
-import { capitalizeFirstLetter, convertStarToWidth } from '../../utils';
+import { capitalizeFirstLetter, convertStarToWidth } from '../../utils/utils';
 
 type OfferPageProps = {
   authorizationStatus: AuthorizationStatus;
@@ -160,7 +160,7 @@ export default function OfferPage({
                 </li>
               </ul>
               {isUserLogged(authorizationStatus) ? (
-                <OfferReviewForm />
+                <ReviewForm />
               ) : (
                 <PostReviewError />
               )}
